BACKPORT, FROMLIST: Support the stable_inodes feature
Reserve the codepoint for EXT4_FEATURE_COMPAT_STABLE_INODES, allow it to
be set and cleared, and teach resize2fs to forbid shrinking the
filesystem if it is set.

This feature will allow the use of encryption policies where the inode
number is included in the IVs (initialization vectors) for encryption,
so data would be corrupted if the inodes were to be renumbered.

For more details, see the kernel patchset:
https://lkml.kernel.org/linux-fsdevel/20191021230355.23136-1-ebiggers@kernel.org/T/#u

(cherry-picked from https://lkml.kernel.org/linux-ext4/20191021233043.36225-1-ebiggers@kernel.org
 and resolved conflicts with this branch not having the fast_commit feature yet.)
Change-Id: Ib100eec48d3fe990ef5df5823b0aaca7b74f039d
Signed-off-by: Eric Biggers <ebiggers@google.com>