resize2fs: fix to respect the environment variable E2FSPROGS_FAKE_TIME
When performing an off-line resize, if an inode's block map needs to
be updated, resize2fs will update the inode's ctime.  In addition, if
inode numbers need to be renumbered due to the file system shrinking
forcing the inode table to be shrunk, any directories which need to be
modified will have their ctime and mtime updated.

If the E2FSPROGS_FAkE_TIME environment variable is set, when the file
system is opened, fs->now will be set to this value, and resize2fs
needs to use it instead of calling time(0) to get their current time.

Cherry-picked from
https://github.com/tytso/e2fsprogs/commit/0288b1fd6909e92e3668dde8f1f6401fdabd1494

Bug: 230874381
Test: debugfs -R "stat /lost+found" /tmp/system_a.img (see 241168624)
Change-Id: I335598c62c9d0ec7a93831bcc0f8f1bc57fb1dfd