Revert "Android: stop suppressing warnings from macOS build"
Unfortunately, the macOS build is not tested either by presubmit or by
local builds.  A macOS build *is* being tested in the upstream GitHub
Actions workflow now; however, that uses the autotools-based build
system, and there can be issues specific to the Android build system.

As a result, removing -Wno-error was not safe yet, and the macOS build
is currently broken in postsubmit.  As there could be multiple issues,
let's restore -Wno-error until I've had a chance to fix the warnings.

Bug: 267448785
Change-Id: I305f73d1f8637477da3d57b6c93037a6e3d9e829