Mostly restore -Werror for macOS build
It turns out the "Can't use getmntent or getmntinfo" warning was the
only warning remaining in the macOS build via the Android build system.
So now that it's fixed, -Wno-error can be removed.

That being said, the upstream CI (GitHub Actions) currently uses
-Wno-error=deprecated-declarations for the macOS build, since it's still
needed for some files (which aren't built by the Android build system).

For now, let's just replace -Wno-error with
-Wno-error=deprecated-declarations to match what the upstream CI uses.

Change-Id: I77f6649b99432ef1d73a0c7e30bbb150c3111b27